furthe refactoring to remove warnings and make compilation more simple and clean
git-svn-id: file:///home/prajwala/workspace/svnrepo/mosland/mos@17 103b0123-5a88-469a-838f-6c959af6773c

+++ libc/include/stddef.h
@@ -18,7 +18,7 @@
 #ifndef _STD_DEF_H_
 #define _STD_DEF_H_
 
-typedef unsigned int size_t ;
-typedef long long ssize_t ;
+typedef long unsigned int size_t;
+typedef long long ssize_t;
 
 #endif
